// front/src/app/model/product.ts
export class Product {
    id?: number;
    code?: string;
    name?: string;
    description?: string;
    image?: string;
    category?: string;
    price?: number;
    quantity?: number;
    internalReference?: string;
    shellId?: number;
    inventoryStatus?: "INSTOCK" | "LOWSTOCK" | "OUTOFSTOCK";
    rating?: number;
    createdAt?: Date;
    updatedAt?: Date;
  }
  